requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.22.34
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
botocore==1.23.34
Brlapi==0.8.3
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.6
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docutils==0.17.1
duplicity==0.8.21
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.14.1
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
future==0.18.2
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.3
importlib-metadata==7.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==2.4
keyring==23.5.0
kubernetes==29.0.0
langchain==0.2.1
langchain-chroma==0.1.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.16
langsmith==0.1.63
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
meson==0.61.2
mmh3==4.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.18.0
openai==1.30.3
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
posthog==3.5.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1_modules==0.4.0
pycairo==1.20.1
pyclipper==1.3.0.post5
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
rapidocr-onnxruntime==1.3.22
regex==2024.5.15
reportlab==3.6.8
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12
systemd-python==234
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
ttystatus==0.38
typer==0.12.3
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.12.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.30.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
